refactor: share diagnostic helper logic
