+++ command.py
@@ -13,6 +13,7 @@ from .db_bootstrap import (
     inspect_lcm_schema_health,
     repair_external_content_fts,
 )
+from .diagnostics import _has_lifecycle_fragmentation, _state_db_path_for_engine
 from .ingest_protection import (
     externalized_payload_stats,
     scan_externalized_payload_integrity,
@@ -31,50 +32,6 @@ from .presets import (
 )
 from .session_patterns import build_session_match_keys, matches_session_pattern
 from .store import build_message_fts_spec
-
-
-def _state_db_path_for_engine(engine) -> Path:
-    hermes_home = getattr(engine, "_hermes_home", "") or ""
-    if hermes_home:
-        resolved = Path(hermes_home).expanduser().resolve() / "state.db"
-        # Check containment within allowed base only when restriction is active
-        env_base = os.environ.get("LCM_HERMES_BASE_DIR")
-        if env_base:
-            allowed_base = Path(env_base).expanduser().resolve()
-            try:
-                resolved.relative_to(allowed_base)
-            except ValueError:
-                raise ValueError(
-                    f"hermes_home {hermes_home} resolves to {resolved} which is not within allowed base {allowed_base}"
-                )
-        return resolved
-    db_path = Path(getattr(engine._store, "db_path", Path.home() / ".hermes" / "lcm.db"))
-    return db_path.parent / "state.db"
-
-
-def _has_lifecycle_fragmentation(stats: dict[str, Any]) -> bool:
-    direct_mismatch_keys = (
-        "lifecycle_current_missing_in_lcm_any",
-        "lifecycle_last_finalized_missing_in_lcm_any",
-        "lifecycle_current_missing_in_state",
-        "lifecycle_last_finalized_missing_in_state",
-        "lcm_message_sessions_missing_in_state",
-        "lcm_node_sessions_missing_in_state",
-    )
-    lifecycle_rows = int(stats.get("lifecycle_rows", 0) or 0)
-    missing_lifecycle_reference_keys = (
-        "message_sessions_without_lifecycle_reference",
-        "node_sessions_without_lifecycle_reference",
-    )
-    return (
-        any(int(stats.get(key, 0) or 0) > 0 for key in direct_mismatch_keys)
-        or (
-            lifecycle_rows > 0
-            and any(int(stats.get(key, 0) or 0) > 0 for key in missing_lifecycle_reference_keys)
-        )
-        or (bool(stats.get("state_db_checked")) and bool(stats.get("state_db_error")))
-    )
-
 
 def _fmt_bool(value: Any) -> str:
     return "yes" if bool(value) else "no"
